Shield
Fixed sword jank, need opinion on attacking system where the sword angles to the mouse and you swing the sword manually by moving the mouse around on the screen. New system is easier to code and doesn't not require a custom attacking animation for each weapon. Also added shield that pulls up on right click.